add plan_validation_test.py
